Apply 1500ms delay to vector (and not vector-2022) only
AFAICT the dancing tab issue in vector-2022 has been fixed as a result
of the great work done in T306229. As a result, we can speed up the
vector-2022 tests by removing the delay.

legacy vector tabs continue to dance so we keep the delay for those
tests for now.