Remove prefers-reduced-motion emulation
Unfortunately, the `transition-duration` in
Ie1c6c1ba7263c232d874263fdae7427a5ec489f6 is causing elements to behave in an
unpredictable manner. Removing the `prefers-reduced-motion` is now more
stable/predictable than using it.

Also, wait until next frame in `fastForwardAnimations`. Note there is still some
flakiness with the visibility of the mobile sidebar that needs to be sorted out
but this should help with the other random stuff.