Add ability to reset the database
The ability to reset the database before a test group runs can be necessary when
using previous branch (e.g. when using the `-b` flag) as that code may assume an
older database schema structure that the database may not have.

Resetting the db will also become critical when tests are introduced that make
changes to the database (e.g. tests that edit an article ). Providing this flag
enables the ability for each test group to be completely isolated from the side
effects that another test group may cause.

Additionally:

* Update database seed file to `database_2022-06-13_12-27-44-0600(MDT).tar.gz`
to fix `crash` messages appearing in the mysql logs.

* Revise `db:save` npm script to enable "slow shutdown" of mysql prior to taking a physical backup. Mysql states this is needed in their docs [1]

* Revise healthcheck logic in docker compose after noticing the healthcheck
reporting that the db was healthy when it wasn't accepting any connections yet.
Relying on a `status` query for the my_wiki should be more reliable.

[1] https://dev.mysql.com/doc/refman/8.0/en/innodb-backup.html