Add clean and stop commands to CLI
The clean command which was previous an npm script (`npm run clean`) and
destroys all containers, images, and volumes related to Pixel has proven to be
useful and should be promoted as part of the CLI API. The stop command which
stops all containers follows similar logic.

Both of these will become critical commands when Pixel is installed as an npm
module.

To destroy all containers, images, volumes related to Pixel:

```
./pixel.js clean
```

To stop all containers related to Pixel:
```
./pixel.js stop
```

Additionally, removes corresponding npm commands since these are now part of the
CLI.