Increase table and desktop breakpoints by 1 pixel
These were leading to too many failures that only happened when the
viewport was either exactly 720px or 1000px wide. These peculiarities
are caused by the way we mix min-width and max-width media queries
(T310536) and are actual problems, but there are bigger fish to fry than
to base the tests on these problematic viewport widths.