Fix "‘node’: No such file or directory" error
This error has occurred in the past, and I thought had been fixed, but it
occurred recently on my new macbook m1 so this is another attempt to fix it. My
guess is that something wrong was happening with the Docker multistage build.
The easiest way to avoid this is to not use a multistage build and instead
install node from nodesource. Hopefully, this resolves this error.