Revise checkout logic
Pixel is responsible for checking out the correct branch for all of its
repos, applying the correct gerrit patch, installing composer
dependencies, and running any database migrations. These need to happen
in the correct order. Before this commit, the maintenance/update.php
script was happening in the middle rather than at the very end.

This commit refactors all of the checkout logic into one class -
MwCheckout and performs the steps in the correct order.

* Replaces `checkoutBranch` and `checkoutPath` scripts with `MwCheckout`
* Adds unit tests with code coverage > 90%
* Moves all unit tests into __tests__ folder which follows the
  convention that Jest uses and that Vector uses.
* Stubs gerrit network requests in a gerritResponses.js file that are
  used by the unit tests.
* Adds an `exec` method to BatchSpawn to enable parallel processing with
  Node's `exec` command.

This ticket can be tested with the following command/gerrit patch which has
dependencies:

```
./pixel.js test -c I8d3af86fdc3daf42441a93fc5b64ebcef37c5fb4
```

Bug: T309739