Revise Dockerfile.mediawiki to leverage docker build cache
Before this commit, Docker's build cache was being busted each time a
change in the `src` directory was made because the `COPY` statement was
placed before the `setupRepos.js` script ran (which clones mediawiki and
all its extensions and skins). This was unnecessary as setupRepos.js
only depends on BatchSpawn.js and repositories.json and led to
unnecessary and time consuming rebuilds of the mediawiki image.

Instead, place the specific files that setupRepos.js needs before it and
defer copying the src folder until after setupRepos.js has executed to
leverage Docker's build cache.

https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#leverage-build-cache