Clean up env variables (#119)
Bernard and I had a long debug session for an odd redirect behavior with his
pixel instance. In short, it was a result of his machine having set the
`MW_SERVER` environment variable from his mediawiki-docker setup. This clobbered
pixel's dependance on that variable being set to localhost:3000.

This commit renames these env variables to include a `PIXEL` prefix and also
cleans up some env variables that are probably not necessary anymore.