Fix Pixel hanging on MacBook Pro M1 by updating backstop/puppeteer/chromium version
This attempts to fix https://phabricator.wikimedia.org/T318445 where Pixel would
frequently hang on MacBook Pro m1 machines

I think the cause of this was that we were using chromium for debian buster
which was only downloading chromium version 89 (the latest chromium version is >
100) which is only compatible with puppeteer 6 (we were using puppeteer 12) [1].

This commit updates the dockerfile to use debian bullseye which will currently
use Chromium 105 [2] and updates to backstop to backstop 6.1.1 which uses
puppeteer 15.5.0.

With these updates, this issue seems fixed!

Additionally:

* Add `init: true` to docker compose visual regression service which gets rid of
zombie processes

* Increase asyncCaptureLimit from 5 to 10 which speeds up the tests by 20
seconds on my machine.

* Add npm i to `./pixel.js update` command so that updates like the one in this
PR can be handled.

[1] https://github.com/puppeteer/puppeteer/blob/ddc567a4b6b94fb888a01521e5f36a9efc5ebf51/website/versioned_docs/version-18.0.5/chromium-support.md#chromium-support