Fix mobile test flakiness
The mobile tests sometimes fail due to differences in color of the toolbar's
text. After a lot of trial and error, I believe this is due to using the
`viewport` selector which seems to have a bug in either backstopjs or puppeteer.
Replacing it with the 'html' selector seems to eliminate this issue.

Also:

* Remove extraneous query string from the `Test (#minerva #mobile #logged-in)`
test.

Bug: T322438