Remove test images as part of `test` command
Each test group generates a folder where it stores its test screenshots. While
pixel automatically removes the reference folder when running the ./pixel.js
reference command, it doesn't remove the test folder when running the ./pixel.js
test command. This can lead to that folder becoming extremely large since it
grows with every execution of the test command. This commit removes that folder
with each invocation of the `test` command.

Bug: T322867