Fix openReportIfNecessary method when report path name is different than the group name (#144)
The desktop-dev group was failing to write to/open the report in the
openReportIfNecessary method because it was looking for a `dektop-dev` folder
instead of a `desktop-development` folder.  Since the report path is
configurable and may differ from the group name, we need to look at the config
in order to determine the report path.

Additionally:

* Refactor the removeTestFolder method to have a relativePath param instead of
passing the entire config.