Make report directory configurable.
This makes it possible to configure the report directory. On the
new server setup, it's then easy to use a new mounted volume where
we can store the reports.

As long as .env is updated, everything will work as before. On the
server we can run: PIXEL_REPORT_DIRECTORY=/mnt/pixel-data ./pixel.js

Or just export that variable.