Add Codex-specific onReady script
The shared one in onReady.js doesn't work because it contains too many
MW-specific things, and errors out if the global mw object doesn't exist
(which it doesn't in the Codex sandbox).

Instead, use a simpler onReady script for Codex, that just prints the
scenario name, waits for idle, and adds a CSS one-liner that disables
animations (this fixes T361528).

Bug: T361528