Add 1 test for Vector and Minerva night mode
The main thing we care about is articles so I think one test
should be sufficient for checking for major issues.

For other pages, I suspect these would be better done as
acceptance criteria to fixing those pages. For example if and
when night theme is supported on talk pages we would add a test
as part of that task.

Bug: T363945